[pretty-print] Use the correct contentType for in-place formatting.
Previously we'd use the contentType of the `uiSourceCode`, which works,
except in cases where there's a binding and the infered contentType
for the local file doesn't match the Content-Type header of the network
request. Instead of this dance, just take the contentType from the
SourceFrame, which already does the right job (because that's also
what we use for syntax highlighting).

Fixed: chromium:1514031
Change-Id: I5fd4bfe5a12a5b625f3c003f5f98c1fbb39cbbf8