Replace TreeModified and EntriesModified Events with a function call
TreeModified Event used to be dispatched when an entry was selected from the menu. Instead, let's directly call treeModified event on the dataProvider.

EntriesModified Event used to modify FlameChartView that the entries were changed and the FlameChart needs rerendering. Since now all the actions appearing in the Context Menu and the ones that change entries, we don't need this action and can instead rerender the FlameChart right after applying the selected action.

Bug: 1469887
Change-Id: Icee4fdf072b94b9f57b06e3b732420106263ff58