[e2e] Improve and reenable cross tool tests.
The Console -> Issues test has been disabled for a long time, because
it was flaky, but it was flaky, because it was testing too much (not
only the transition between the panels, but also specific data within
the Issues tab afterwards, which wasn't relevant to this test).

Drive-by-fix: Also cleanup and simplify the test setup a bit and
remove some unnecessary gymnastics from the Performance -> Sources
tests.

Bug: chromium:1375161
Change-Id: I871fb7d1960ec7a8d403260e31bf50b4d899c0c2
Fixed: chromium:1342045