Added aria-label "Learn more" to recorder#selector links in recorder tab for screen reader
Narrator only announces "link" when focusing on the recorder#selector links in Recorder.  The aria-label "Learn more" was added so that it Narrartor announces "learn more".  This is similar to how it announces for other learn more links.

Added aria-label learn more to learn more links in recorder tab

Before:
https://imgur.com/a/pUc1EDn

After:
https://imgur.com/a/cFG0lqI

Bug: 1516957
Change-Id: Ibf697867eff92c40b341b09d5ab55f795d609c6d