Fix Screen Reader does not convey group info for Ignore List setting
In the Ignore List setting, the elements were constructed as a flat list but presented as groups visually.
This CL change the elements hierarchy to be two groups: "General Exclusion Rule" and "Custom Exclusion Rule" so the screen reader can properly interpret it.
This does not change the UI, but only changes the elements hierarchy.

https://imgur.com/a/odymkPw

Bug: 1517015
Change-Id: Ie6194230c62651cb91e8a0abea29ebaa49323db4