Display built-in constructor names for heap snapshot
I recently spoke with some JavaScript developers who thought that
"(closure)" in the heap snapshot viewer represented the environment in
which a function runs, rather than the function instance itself. The
column in which this name appears is titled "Constructor", so I think
that it would be clearer to name this category "Function", since that is
indeed the constructor for the objects listed within the category. The
name Function is more consistent with the names Object and Array, which
also appear in the same column.

For consistency, this change also updates "(regexp)" to "RegExp". The
other names "(string)", "(symbol)", and "(number)" should stay as they
are because they represent unboxed primitive values.

Bug: 1517353
Change-Id: I17ad2e745174adc73efe51b178179d5fe1e0fd08