[CSSProperty] Fix formatStyle formatting `var()` params with additional newline
`def` is assigned when[1]:
* It is a variable definition.
* It is an `@` rule.

Since formatting nesting is not in scope of this
`formatStyle` function as CM5 parser is not equipped for that;
the `@` rule case is irrelevant (though it doesn't break
in any case).

We're aiming to migrate to CM6 but there is a blocking issue
there (crbug.com/1514071); so until then, for M122; I've decided
to fix this issue while using legacy parser.

[1]: https://github.com/codemirror/legacy-modes/blob/main/mode/css.js

Fixed: 1518839
Change-Id: I7401d4ccf866a50c31851fdb1c6e68c54c56cf30