[cleanup] Remove ResourceType from ContentData class
We only required it to determine if a ContentData is textual content
or not. The MIME type conveys the same information. To this end this
CL introduces a hard-coded list of MIME types that have textual content
besides "text/*".

This is somewhat cleaner, as ContentData now bundles the minimal
information to pass around bytes in the frontend: The raw data,
the mime type and a character encoding.

R=wolfi@chromium.org

Bug: 1509336
Change-Id: I70061a263c0a18ebc7c435bdb611164ec9363e84