[network] Use ContentData for intercepted response body
This CL changes the responseBody getter for InterceptedRequest to
return the new ContentData. We also parse out the mime type and
charset from the response header received in the Fetch.requestPaused
or Network.responseReceived event.

This fixes one bug where we unconditionally decode intercepted
content as utf-8 instead of applying the right charset.

As a follow-up we should also store the overridden response body
with the original charset on disk, currently we also store it
as UTF-8, which then later causes a missmatch in the backend as
the backend looks at the Content-Type header to decode the overriden
content.

R=wolfi@chromium.org

Bug: 1511581,1509336
Change-Id: Iafa256d42c09b72502624596131beb82bd567a64