[icon] Don't make `Icon` depend on `ThemeSupport`.
This refines the fix from https://crrev.com/c/5199087 to avoid
introducing the dependency from `Icon` on `ThemeSupport`. We don't
need the actual theme support, but only the logic to insert a
`<style>` tag and that we can just do in-line. The dependency
that was added before broke the component server in many cases
and even caused issues with the internal roll.

Bug: chromium:1518436, chromium:1421730, chromium:1514825
Change-Id: I73dc904e76cdc4c3095e69ea7fab928fb5ec075d