DO NOT LAND: (MESSY) Only keep FORCED_X warning if it was JS-invoked
Commented out: an overly complex mechanism that uses the tree to see
if we're in a script even

Not-commented out: A way simpler version.. Is there a stackTrace attached
to the style/layout event?

I don't know that these two are functionally equivalent, but they were in
the 5 cases i checked. Sometimes stacktrace is an empty obj/array but..
that counts as "active".

Change-Id: I15f6643fb726e1404cf1babfcf5c0756327fcf6e