[GN] Fix missing/incorrect build dependencies for `meta` targets.
This was noticed with https://crrev.com/c/5210758 (and caused that
CL to be reverted): The `meta` targets only transitively depend on
their imports (most of the time), which means that if one of their
imports change, but this change does not cause the `bundle` target
on which the `meta` target depends to change its output, ninja will
not rebuild the `meta` output. This causes issues specifically with
changing `enum` to `const enum`, as the *-meta.ts files don't get
rebuild correctly.

Fixed: chromium:1520414
Also-by: szuend@chromium.org
Bug: b:319814509, chromium:1172300, chromium:1167717
Change-Id: I436b349f3d99051848c7e7dbf63b35b5926d1448