[network] Emit text/event-stream events for non-EventSource requests
This CL hooks up response body streaming with a "server sent events"
format parser. This is required for requests of type "fetch" or "XHR"
that use the text/event-stream mime type.

In this case we don't get the CDP events themselves but have to do
the parsing ourselves.

We move all this logic into a separate class that also stores the
already received messages.

R=wolfi@chromium.org

Bug: 1509336
Change-Id: Iae951e25e8c067a259850e668964ba9f9c3bc4c2