[issues] Remove statistics for "Learn more" links.
With the new VE logging in place we should be able to get much better
data than with the previous logging. Currently the VE logging doesn't
yet offer the same level of detail, but so far it also didn't seem
necessary to have the current level of detail, and we can always
enrich the VE logging with more detail if that turns out to be
necessary.

Fixed: chromium:1520446
Change-Id: Iffc905bc7a9dcbc87d0034b369d12337cc8cd708