[text_utils] Add StreamingContentData
This CL adds a new StreamingContentData class that allows for binary
data to be received piecemeal and notify consumers when new data is
available.

StreamingContentData constructs the concatenated ContentData in a
lazy manor only when requested.

StreamingContentData can wrap an existing ContentData, but does not
allow adding new data to it (due to a technical limitation of JS).

Drive-by: Expose 'charset' on ContentData as a public readonly
property.

R=pfaffe@chromium.org

Bug: 1509336
Change-Id: Iec77a163e44656a4eeac2213a1132fc5888180b9