Add StreamingContentProvider and implement it on NetworkRequest
This CL adds an extension to the ContentProvider interface that can
proivde StreamingContentData. The CL also implements it for
NetworkRequest.

Note that the "streamResourceContent" CDP method only works for
in-flight requests. For already finished requests we use the
normal "contentData()" method to get the full content and wrap
it into a "StreamingContentData" that won't get updates.

Since "ServerSentEvents.ts" already relies on response body streaming
we re-wire it to use the new StreamingContentData instead.

R=pfaffe@chromium.org

Bug: 1509336
Change-Id: Idb036dbac81d2cc28540f9819674e5b7acf008c6