Simplify EntriesFilter by removing Undo actions
Filter actions were initially designed to be separated into 'Apply' and 'Undo' actions because the plan was to be able to undo specific actions and for that we needed to map an action to a specific undo actions.

The final design only allows to either reset the whole trace or reset children of a specific entry. For that we don't need a separation into different actions and having different ttypes of actions only adds complexity.
Let's get rid of all the code associated with Undo actions and move 'reset trace' and 'reset children' into FilterActions enum.

Bug: 1469887
Change-Id: Idc8499db358e681db76a70c6b83b96281a923d66