Improve retained size computation for WeakMap keys
When an object is stored as a value in a WeakMap, and there are no other
references to that object, the corresponding key and the WeakMap itself
could both reasonably claim that they retain the size of the value
object, since removing either the key or the WeakMap would be sufficient
to remove the value. However, DevTools currently finds the common
dominator of the key and the WeakMap and assigns the retained size to
that common dominator instead. This is less useful; there might be many
intermediate objects which would benefit from better visibility if they
were assigned a higher retained size.

To generate the most correct output, we would need to move away from
using a dominator tree for computing retained sizes, and instead invent
a new notion similar to dominators but where each node can have multiple
immediate dominators, since there are multiple nodes which all must be
present to keep a selected object alive. However, this would introduce a
lot of unwelcome complexity, and I haven't though of a reasonably fast
way to implement calculateClassesRetainedSize when the new dominator-ish
relationships form a non-tree DAG. Instead, this CL implements a much
simpler change which should still be beneficial in many cases: only
consider the edge from the key, not the edge from the WeakMap, when
computing dominators.

Bug: 40818377
Change-Id: Ie6a21eca5d0517a644809d98b2b357e0105a9775