Infra for RPP extensibility based on DevTools extension API
This adds the first changes to the DevTools extension API to
support extensibility in the performance panel. This only includes the
bits in the devtools extension layer to support adding a new track to
the Performance Panel using a DevTools extension.

Note that the UI changes are not done here and as such nothing is
displayed. This is done as a follow up of this CL.

Bug: none
Change-Id: If259c1c2182dd8b6a2eda24029a08e4da52944d5