[text] Add SafeContentProvider to migrate away from DeferredContent
We'd like to replace "DeferredContent" with "ContentData" since the
latter carries MIME type and charset info with it and will always
correctly decode text content from base64.

Since DeferredContent is returned by a widely used interface method,
we extend the "ContentProvider" interface with a "SafeContentProvider"
interface that returns a "ContentData". This way, we can first update
all providers piecemeal and then switch over consumers one by one.

The CL also introduces a variant of "SafeStaticContentProvider" as
well as a small helper to search in "ContentDataOrError".

R=alexrudenko@chromium.org

Bug: 41481944
Change-Id: I6a945109e044c7452f4918c6de4ca5f6c023ce54