[presubmit][deps] Format touched files fully when we suspect large diffs
'git cl format', by default, only formats the changed lines for each
file. It does so by forwarding the changed lines from git to
clang-format on the command line.

For large diffs, this results in very large command line invocations
that run into the limit on windows.

As a heuristic, we should just format all changed files fully if we
suspect this might be the case.

Note that previously we skipped formatting alltogether when e.g.
node_modules was changed.

R=alexrudenko@chromium.org, ergunsh@chromium.org

Bug: 40272289
Change-Id: Ic7648a0ad81f6e27f84e07c3497542a61acd4d3a