Improve predictability of interaction tests
1) RPP interaction test timeouts should be disabled when in DEBUG_TEST
If using `npm run debug-interactionstest`, any timeout overrides
shouldn't interfere with debugging.  The `this.timeout() !== 0`
pattern used elsewhere is reused here.

2) Many HTML documents used in interaction/e2e tests were in quirks
mode. This results in very slight differences [0] like unexpected
table styling. We apply a doctype to keep the component pages
consistent with how the frontend is delivered.

3) Hide component docs UI during interaction tests to prevent accidental
click()s on it.

[0] https://quirks.spec.whatwg.org/

Bug: chromium:325316147
Change-Id: I7a9defa4006382c130e384fe1bcc671bb7f00505