[test] Audit and cleanup uses of assertNotNullOrUndefined().
Now that `e!` is allowed in tests, we can simplify the code in various
places and avoid the noisy `assertNotNullOrUndefined(e)` to silence tsc.
We still want to do this carefully to avoid making it more difficult to
investigate failures, so we only use `e!` in places where a subsequent
property access will fail in case of `e` being `null` or `undefined`.
This way we can simplify the test code in many cases, making it more
friendly to the human eye, while retaining the important property of
failing early in tests and not propagating errors until later in the
test execution.

Doc: http://go/chrome-devtools:eslint-pain-points-design
Bug: b:319814509
Change-Id: Ie1d2af78aad56273ff3e36ccabaef6a1cb8130d4