[ServiceWorker] Add transferred size in the network panel
This CL adds the network transferred size information to the network
panel for the service worker static routing API [1]. When the service
worker has routers, and incoming requsts matched to the router which
lets requests go directly to the network, currently DevTools doesn't
show the network transferred size, while developers can see it on other
regular network requests.

This CL add the info to the entry in the network panel when developers
hover the cursor to the size column.

Here is the chromium side CL: crrev.com/c/5297354

[1] https://groups.google.com/a/chromium.org/g/blink-dev/c/-SEMMttEqz4/m/ogpF3VZsAAAJ

Bug: 325120866
Change-Id: I392f9d7743abe10f0a0738d9a630c5f892052310