[cleanup] Remove obsolete SCSS formatting.
SASS files (and in particular those using SCSS syntax) are considered
authored files, and we should not try to pretty-print them, but leave
them as the developer authored it. Also remove the dedicated tests,
keeping those that are useful for CSS, but removing the SASS specific
ones, which only passed by coincidence anyways, since our CSS formatter
is sufficiently generic.

Bug: b:323795674
Change-Id: Ie63ce0bccdf7035bd696df81b832e9d4b7ca247e