[cleanup] Remove unnecessary `defineComponent()` helper.
The helper was originally introduced with https://crrev.com/c/2878729 in
order to prevent accidental duplicate registrations, but that is not an
actual problem in practice. On the flip side the helper adds unnecessary
noise (both in terms of readability/familiarity and in terms of
additional pragmas to disable the relevant ESLint rule). And
additionally Open Web Component tooling[^1] doesn't understand the
helper, and cannot infer Web Component information for our code
correctly.

Given the downsides (and basically no upsides) and the fact that we are
actively exploring the idea of adopting more of open-wc, we'll drop the
helper and its eslint here.

[^1]: https://open-wc.org/guides/tools/web-component-devtools/

Doc: http://go/chrome-devtools:eslint-pain-points-design
Bug: b:319814509, chromium:1205825
Change-Id: If888be52c94d97f09662543fbf604f8dfe1eb575