[ServiceWorker] Add transferred size in the network panel
This CL adds the network transferred size information to the network
panel for the service worker static routing API [1]. When the service
worker has routers, and incoming requsts matched to the router which
lets requests go directly to the network, currently DevTools doesn't
show the network transferred size, while developers can see it on other
regular network requests.

This CL add the info to the entry in the network panel when developers
hover the cursor to the size column.

Here is the chromium side CL: crrev.com/c/5297354

Other router sources:
For `cache`, we don't need to show the transferred size, as there is no
network transfer. This is consistent with the diskcache case.
For `fetch-event`, another entry in the network panel, from the service
worker to the network, should show the info as the regular SW fetch
handler does. The current behavior is not consisntent with that, but
that should be supported in a separate CL.
For `race-network-and-fetch-handler`, TBD. Maybe we should have a plan
after supporting `fetch-event`.

[1] https://groups.google.com/a/chromium.org/g/blink-dev/c/-SEMMttEqz4/m/ogpF3VZsAAAJ

Bug: 325120866
Change-Id: I392f9d7743abe10f0a0738d9a630c5f892052310