[test] Audit and cleanup several uses of assertShadowRoot().
Now that `e.shadowRoot!` is allowed in tests, we can simplify the code
in various places and avoid the noisy `assertShadowRoot(e.shadowRoot)`
to silence tsc.

We do this very carefully only for places where `e.shadowRoot!` is
(immediately) followed by a property access to not hide or delay
bugs.

Doc: http://go/chrome-devtools:eslint-pain-points-design
Bug: b:319814509
Change-Id: I7d712863bb3ba0c325c82889fbfe048bdf0f4f78