RPP: Disable entry link when the entry is outside of the current breadcrumb
If an entry the link is for is outside of the current breadcrumb, disable the link and add a disclaimer for this.
If the MiniMap bounds get updated, dispatch an event to rebuild the summary in case an entry with a link was previously selected.

Also added a test to navigate to an initator and back to the previous entry from it.

Bug: 326214222
Change-Id: I1c453807db1f55d779d07cb349fa99662540edbc