Reduce time spent copying sample data in buildHierarchy
The RendererHandler was doing a lot of copying to reconstruct Maps of
sampler data, for every thread in the trace. This data doesn't need to
be copied at all, so I extracted the call to it out of the loop.

Additionally, I changed the SampleHandler data to not copy its contents
to new Maps, so now instead of copying this huge data once every thread,
there are zero copies.

For a 150MB trace (~8s trace time), this copying took ~1s on my M1 mac.

Traces are available here:
https://gist.github.com/connorjclark/c972b4e85284a9199f91e1da7abf91b0

Bug: none
Change-Id: Ifcc6283aa1f149acb6813535b1c3d2dbc2671a56