Reduce time spent copying sample data in buildHierarchy
The RendererHandler was doing a lot of copying to reconstruct Maps of
sampler data, for every thread in the trace. This data doesn't need to
be copied at all, so I extracted the call to it out of the loop.

I attempted to change the SampleHandler data to not copy its contents
to new Maps, but 2 tests failed. I suspect the tests just need to be
rewritten slightly, but as this is now just a single ~8ms call, I
did not look further. So there remains one copy, outside the loop.

For a 150MB trace (~11s trace time, some navs), this copying took
~1s on my M1 mac.

Traces are available here:
https://gist.github.com/connorjclark/c972b4e85284a9199f91e1da7abf91b0

Bug: none
Change-Id: Ifcc6283aa1f149acb6813535b1c3d2dbc2671a56