[cleanup] Remove `Platform.DCHECK`.
This was originally introduced with https://crrev.com/c/2894390 but we
never started using it consistently. It also comes with the downside
that unlike in Chromium, for DevTools debug builds aren't used as
extensively and frequently, so it's better to stick to either throwing
an Error or console.assert() for the few cases where we used the
DCHECK() function in the past, and thereby ensure that we are notified
of the problems early on.

Bug: none
Change-Id: I90bd465d088be5530530027ea1dbd0ec2b74b12b