[deps] Set LPAC ACLs on windows for CfT binary
In some cases, windows bots don't have permission to run the CfT
binary due to missing ACLs.

Chromium tests usually set this ACL on the out/ directory when
building. For us it's enough to set up the ACL when we download the
CfT binary.

Manually verified on a Windows cloudtop that ACLs are correctly in
place after downloading a fresh binary.

R=mathias@chromium.org

Fixed: 324929074
Change-Id: Ieb5cb18ff883c1ee305be9738c725c3ebca00262