[RPP] Refactoror initiators
Instead of creating 4 arrays with startime, endtime and levels, pass indexes of the initiator and event to the Flamechart.
This is needed because in the future we need to be able to draw the initiator arrow from the middle of the entry and to do that, we need to know its' index in the Flamechart.

In the next cl, I will combine the logic to draw circles around arrows with drawing the flow events and draw the arrows directly to/from the circle.

Bug: 326916488
Change-Id: Ib8fb6f8dafb0eff3b1ee04323135b1729cac8e22