[Flame Chart] Add UI implementation for track configuration
This CL adds the UI for move up/down, and hide a track.

When you hover a track, it will show an edit button.
If you click the edit button, you will enter the editing mode, and you can move
a track up/down, or hide it. And there is a save button to quit the editing mode

Bug: 311439339
Change-Id: I8bf6a85a4747bcdbb3106dcddfc39c25e5036a71