Add descriptive tooltips for some types in the heap snapshot viewer
Any time the user hovers over an object in a heap snapshot, DevTools
presents a tooltip. For some object types, this tooltip includes
detailed information, such as the list of properties on a JavaScript
Object. For other types, the tooltip says simply "Preview is not
available". This change adds short descriptive text for a few of those
types which show up frequently, plus a link to a longer description in
the DevTools documentation.

Bug: 41490331
Change-Id: I6ac8d0be1b597bf6d02b7c01b7ba1b3573ec8d3a