Fix stack preview not getting rendered correctly in initiator
The reason this happened is, in crrev.com/c/5172554 we've changed
the layout logic from the links being able to grow the width of
the container to them staying inside the container.

However, some usages expect that to grow the container. This is a
quick solution to parameterize that behavior and only use it
in console view message.

Fixed: chromium:327665602
Change-Id: I61891883de7a13f40f53b73d8bce9ce90c15917a