ThemeSupport: Fix misconfigured caching on getComputedValue
The caching was keyed off of the return value of a getComputedStyle()
call, however per the [NewObject] IDL, that object is a one-off.
This slip meant we'd always have a cache miss, but perhaps more
importantly, we'd retain a new Map() on every getComputedValue()
invocation. (Perf Panel calls it ~100x per rAF)

Also, the elem.getComputedStyle invocation moves to be behind cache protection. Lastly, some renames for clarity.

The result: fewer calls to getComputedStyle; a perf win. And fewer Map
objects retained, making GC happier.

Bug: 328733321
Change-Id: I9cbce7aa28b15b14ab8147200222f8bc7f1328f1