Update the Galaxy Fold device emulation target.
The device is not maintained anymore so it's probably better
for developers to test on a newer target. This patch also
make sure that the the hinge is correctly setup so developers
can try their websites with Viewport Segments and Posture
correctly set up. The hinge width/height of 0 (depending the
orientation) is following what Android OS returns (unfortunately).

Bug: 40113439
Change-Id: Ifd6b4761e14ad536e103101c93105360a771c29c