[network] Fix HTML preview
ContentData already takes care of properly encoding data URLs when it's
needed. The HTML preview view doesn't have to encode it again.

This CL also moves the creation of the HTML preview into a static
factory method, so callers don't actualy have to worry about this and
just pass a ContentData.

R=bmeurer@chromium.org

Fixed: 328471611
Change-Id: I2ea32d85157599ec695147a56f2e3ac59f6191db