Remove duplication for timeline styles in perf panel
As part of migrating the new engine to the perf panel we included UI
utilities that aid in defining styles for events in the timeline.
Because the perf panel already had similar utilities and objects for the
same purpose we ended up with a bunch of duplicated artifacts.

This CL migrates the perf panel code to use the most recently introduced
utils and deletes the old ones for simplicity. As a side win the
humongous TimelineUIUtils file is a little bit smaller and we have
better type checks.

Bug: 40287735
Change-Id: I70d1f58958ad4f8a2b5c05a5b91101d0c8e5314f