[RPP] Make Entries Filter a Singleton
Let's make Entries Filter a Singleton because we will only ever need 1 instance since every entry has unique Synthetic IDs and EntriesFilter is manipulating Synthetic Ids to keep track of hidden entries.

We are doing this because this allows us to delete multiple function call chains and we will need to create longer ones to start using EntriesFilter in saving annotations.
Additionally, this allows to access the EntriesFilter from TimelineUiUtils and add "event hidden" description to the event link.

Bug: 1469887
Change-Id: Iedda089de58f13237d869023676bf52bb3ac0f7d