[unittest] Also set LPAC ACLs on win when running unit tests
The CL https://crrev.com/c/5330806 sets the ACL when downloading the
CfT binary during "gclient sync". This is not sufficient for our
CI bots as the unit test suite is run on swarming where the relevant
files are just copied over.

To fix these sporadic failures we'll also set the ACL on windows
as part of `run_unittests.py`.

R=pfaffe@chromium.org

No-Tree-Checks: true
Bug: 324929074
Change-Id: I200b78a2ed1f31ffa24a696266e508a4deabc808