Update description for `serviceWorkerRouter` message
For l10n messages:

* We're replacing markdown content with `MARKDOWN_SNIPPET_{X}`.
* We're deduplicating the messages when the contents of the message are the same by adding a meaning field as the description.

However, for the `serviceWorker` and `serviceWorkerRouter` messages
both the message and the description is the same; thus they have
the same id in translation console. This is causing collisions
where the same id is used by messages with different placeholders.

This CL simply updates the description of the `serviceWorkerRouter`
so that they will have different `meaning` fields and will result in different message ids; removing the collision.

Bug: none
Change-Id: Id3297a0f02cc371a41f0af99ac112600cddf795b