Avoid shallow container copies in most trace handlers
Handlers that depend on others use the `.data()` method. The
convention has been to always shallow copy the containers in
these methods. Since what `.data()` returns is typically data stored
in global containers, this copying was important for correct behavior. It is what allows callers to assume the result
of the trace processor will not change, enabling a cache to work
(see TraceLoader.ts), and in general allowing the result to remain
valid even after the processor is reused or torn down.

However, callers of `.data()` in dependent handlers must be aware
of this behavior, else they risk a big performance cost if they call
`.data()` too much. This was the cause of a large performance hit that
I recently fixed [1].

This CL avoids shallow copying in `.data()` and instead does a copy
at the end of the trace processor execution. This allows handlers to
use others' `.data()` without concern of performance overhead, and
protects against similar performance regressions in the future.

The two handlers that still contain some shallow copying should be
straight-forward to resolve, but I wanted to avoid making any changes
to tests or implementation for the majority of this refactor. See:
LayoutShiftsHandler and UserTimingsHandler.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5318630

Bug: 41484172
Change-Id: I5a404c3b14d5ebea627f96d141e363242a65c860