Support bigger heap snapshots
I recently received an email about a heap snapshot that couldn't load in
DevTools because it had 227 million edges. This is clearly pathological
behavior, but the whole reason they wanted to open the snapshot is to
investigate that pathological behavior, so I think that DevTools should
make every effort to support such a large snapshot. We can allow loading
that snapshot if we represent containmentEdges using a new type which
might internally use multiple Uint32Array instances. This makes accesses
into the containmentEdges array a little uglier, but I think that it's
overall a worthwhile change.

Change-Id: I039ae6b67c43a13b60b35d281f0d04f27bf183c1