[theme_support] Fix race condition when opening with Cmd+Shift+C.
https://crrev.com/c/5307616 introduced a very subtle timing issue by
waiting for the `theme/colors.css` file to load when starting the
DevTools app. This innocent looking `await` change causes the
keyboard event handling (which is in a completely different place) to
be unable to toggle the select tool.

This CL reverts the one line related to the `await` and instead puts
the call to `applyTheme()` into the `onload` handler, and thereby
fixes the timing issue.

Fixed: b:330587444
Bug: b:41495790
Change-Id: I8e11560ea30dc6d682e9724edb1c30771eb99180