Reland "[console] Implement basic Error.cause rendering"
This is a reland of commit cf229f3b2820ceda5fcceda045b909ab32bc9660

The reland includes a fix for a CSP e2e test. Rendering stack traces
has always been asynchronous, but the Error.cause CL makes it worse:
We create the 'div' first and then populate it incrementally and
asynchronously with stack traces.

The CSP e2e test only waited for the 'div' and checked the
div's textContent while we were still building the stack trace.
The fix is to also wait for the stack trace to be present.

Original change's description:
> [console] Implement basic Error.cause rendering
>
> This CL adds rudimentary support for Error.cause similar to what
> firefox is doing. We'll recursively walk the 'cause' chain and print
> each errors stack trace with a "Caused by: " prefix.
>
> This CL supports cause objects that are either "Error" objects
> themselves or plain strings. A possible improvement could be to
> support arbitrary JS objects and call `toString` on them.
>
> We add a simple e2e test to cover the basic user story, aka
> "Error.cause works".
>
> R=pfaffe@chromium.org
>
> Fixed: 40182832
> Change-Id: I7082e7144b1e6cd9dc20d0409db1e25279780a1d
> Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5382167
> Commit-Queue: Simon Zünd <szuend@chromium.org>
> Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>

Fixed: 40182832
Change-Id: I9f9dae16d644cab54794f446d56006861ff322f0