Sort unreachable nodes to the end in the Retainers pane
V8 attempts to avoid including unreachable objects in heap snapshots,
but they can still be present in some rare cases. I recently opened such
a snapshot and found that the Retainers pane was not ideal because it
sorted unreachable objects to the beginning rather than the end. This
change causes unreachable objects to be sorted after reachable ones,
like we already do when the user has ignored any retainers.

Bug: 327337527
Change-Id: I0a1fc87a318236644ef904e594d2b0cdfe166edf