[Flame chart] Fix the bug that entries are not shown after reorder the tracks
Before the |visibleLevelOffsets| is always sorted, so we used bisect to find the
first visible level, and then traverse till the last visible level, then break.
But after the reordering, the |visibleLevelOffsets| is no loner sorted, so we
need to check each level now. Since the max level number won't be too bigger, so
the time complexity increase should be acceptable.

Fixed: 330708252
Change-Id: Ied18540a7664db1c04e5e58501845f3f95b3cf2b