[LayoutShift] Fix injected iframe event filter
Follow-up to:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5384002

Lighthouse issue:
https://github.com/GoogleChrome/lighthouse/issues/15869

Side-note: The iframe root cause check casts a very wide net for events
that could have caused an iframe to be added. I think we should consider
making the filter a bit more strict so we aren't calling getNode on
every node that had a style update.

Bug: None
Change-Id: I68d51ac6b3ed8910ef0b032f885608baf04594b9