[cleanup] Move separated UI utils back into UIUtils.ts.
During the ESM migration, some UI utils were moved out of UIUtils.ts (as
part of https://crrev.com/c/1967134), but there's no advantage/reason
for why this is beneficial. And now it's blocking the removal of the
`__widget` and `__widgetCounter` properties, because some of the
separated UI utils implicitly depend on Widget.ts, but without properly
importing anything from Widget.ts, and they get away with it because of
the use of `any`.

Bug: 332815658
Change-Id: I52742d10d6b0e4aba02b6a055b5c55c51c920023