Don't use empty frames as mainFrame candidates
This CL fixes a rare case where we have a frame with no parent, no name
and no URL being used as the main frame, which is incorrect.

The fix is to only consider a frame a candidate as the main frame if it
has no parent (existing behaviour) and if it has a URL (new behaviour
introduced in this CL).

Fixed:41494290
Change-Id: I1190e1cd7b657fa1f7dc3cf44706fe6662a9353d