[text_editor] Make auto bracket/tag closing configurable.
So far there was no way for the developers to turn off the automatic
closing of open brackets/tags, but we enforced this for everyone. This
CL adds a new setting "Auto closing brackets", just like in VSCode,
which toggles the automatic closing of brackets (and HTML tags) the same
way that it works in VSCode.

Fixed: 40865010
Screenshot: https://imgur.com/Zk1B58e
Change-Id: I447ecc79c5de0787092c760d762adca1224fa328