Remove TimelineModel#tracks
None of them are ever constructed any more.

Note that this CL does not remove the private instance variable
`tracksInternal`, which will need to be removed. However, I am trying to
land this CL to see what happens with layout tests (I expect breakages).
Once the public API removal (this CL) has landed, it will then be safe
to remove the internal variable usage in a follow-up CL.

Bug: 40287735
Change-Id: I0649c5217135d158d4e51509262236ad0a82082a