RPP: add helpers to expose traceEngine in layout tests
There are a couple of layout tests that are much easier to migrate if we
expose the data to them from the trace engine. These are the axe / a11y tests
(which need the trace engine data to properly load up DevTools to run
the tests), along with a test that ensures when DevTools is open we
don't trigger style recalcs [2].

This methods are strictly to enable the layout tests to have no
dependencies on the legacy TimelineModel so that we can be unblocked
with its removal.

1: third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/performance/performance-pane-a11y-test.js
2: third_party/blink/web_tests/http/tests/devtools/elements/styles-4/styles-should-not-force-sync-style-recalc.js

Bug: 40270061
Change-Id: I16fd8a38437965cb0ab017820961cd3358ba5deb