update socket.io-parser from 4.2.1 to 4.2.4
This change is to upgrade the socket.io-parser package from 4.2.1 to 4.2.4 to solve the security issue tracked in crbug.com/40275603.

The update was initially applied in this commit on 7/13/2023: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4682214. But got reverted in this commit on 7/18/2023: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4693881.

Bug:40275603
Change-Id: I6b67df83f77f8fbae09acf36c8f72de87ee4949c