npm package security updates
Change-Id: I2113ee968ec988dd36c8d4bcb5a702ada2ee9329

This fix is to upgrade the socket.io-parser package from 4.2.1 to 4.2.4.

Initial fix was made in this commit on 7/13/2023: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4682214. The fix was partially overwritten in this commit on 7/18/2023: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4693881.

We got a mixed results for the socket.io-parser package after the overwritten: 1) the package version shows 4.2.4 but 2) resolved as 4.2.1. Because of this mismatch, when we rerun `npm run audit fix`, no update will happen on the socket.io-parser package.

In this change, I manually fixed the version mismatch, then reran `npm run audit fix --audit-level=critical` to update the socket.io-parser package from 4.2.1 to 4.2.4.

Bug: 40275603